defineFirebaseuiAuth として切り出した
